% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_volume.R
\name{delete_volume_cache}
\alias{delete_volume_cache}
\title{Deletes the entire search volume cache file}
\usage{
delete_volume_cache(cache_path = "mminer-cache")
}
\arguments{
\item{cache_path}{Path to the directory with the cache. Defaults to the
'mminer-cache' in the current working directory.}
}
\value{
TRUE if file exists, FALSE if it doesn't.
}
\description{
Deletes the entire search volume cache file
}
