% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_volume.R
\name{get_volume_cache}
\alias{get_volume_cache}
\title{Reads search volume data from cache}
\usage{
get_volume_cache(query = NULL, cache_path = "mminer-cache")
}
\arguments{
\item{query}{An optional character vector of queries to read from cache.}

\item{cache_path}{Path to the directory with the cache. Defaults to the
'mminer-cache' in the current working directory.}
}
\value{
A tibble (data frame) with search volume data from cache. If cache
  file doesn't exist, returns NULL. If the _query_ argument is provided,
  returns the matching queries only.
}
\description{
Reads search volume data from cache
}
