% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_volume.R
\name{get_volume_cost}
\alias{get_volume_cost}
\title{Calculates costs for a search volume API call}
\usage{
get_volume_cost(query, cache_path = "mminer-cache")
}
\arguments{
\item{query}{Queries to retrieve volume data for as a character vector.}

\item{cache_path}{Path to the directory with the cache. Defaults to the
'mminer-cache' in the current working directory.}
}
\value{
None.
}
\description{
Prints a message how many queries are already cached and the count of the
rest multiplies by 3 credits.
}
