% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_volume.R
\name{get_volume_data}
\alias{get_volume_data}
\title{Retrieves search volume data of the provided search queries}
\usage{
get_volume_data(
  query,
  lang = NULL,
  api_key = Sys.getenv("MARKETING_MINER_API_KEY"),
  cache = TRUE,
  cache_path = "mminer-cache"
)
}
\arguments{
\item{query}{Queries to retrieve volume data for as a character vector.}

\item{lang}{A language (actually a country) abbreviation as a character
vector of length 1. For instance "us", "gb", "cs", "sk" or "pl".}

\item{api_key}{Your Marketing Miner API key. Defaults to the environment
variable named "MARKETING_MINER_API_KEY".}

\item{cache}{A logical value stating whether the result should be cached.
Dafaults to "mminer-cache" in working directory.}

\item{cache_path}{Path to the directory with the cache. Defaults to the
'mminer-cache' in the current working directory.}
}
\value{
A tibble (data frame) with search volume data for the provided
  queries. Rows are ordered the same as the provided queries.
}
\description{
Calls Marketing Miner Profilers API endpoint '/keywords/search-volume-data'.
If a cache exists, the function reads it first and calls the API only for
queries that are not cached.
}
\details{
API calls are charged according to
\url{https://help.marketingminer.com/en/article/profilers-api/}. In the time
of writing it was 3 credits per keyword.
}
