% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mminer-package.R
\docType{package}
\name{mminer-package}
\alias{mminer}
\alias{mminer-package}
\title{mminer: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MarekProkop/mminer}
  \item Report bugs at \url{https://github.com/MarekProkop/mminer/issues}
}

}
\keyword{internal}
