% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_volume.R
\name{prune_volume_cache}
\alias{prune_volume_cache}
\title{From a search volume cache removes queires fetched from API before a specific
date and time}
\usage{
prune_volume_cache(cache_path = "mminer-cache", older_than = NULL)
}
\arguments{
\item{cache_path}{Path to the directory with the cache. Defaults to the
'mminer-cache' in the current working directory.}

\item{older_than}{Date-time. Queries with the older time stamps will be
removed from cache.}
}
\value{
Number of removed queries.
}
\description{
From a search volume cache removes queires fetched from API before a specific
date and time
}
