% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mminer-package.R
\docType{package}
\name{mminer-package}
\alias{mminer}
\alias{mminer-package}
\title{mminer: Marketing Miner Profilers API wrapper in R}
\description{
The mminer package is a simple wrapper for Marketing Miner Profilers API. It fetches search volume data and cache them locally to save credits when necessary.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MarekProkop/mminer}
  \item Report bugs at \url{https://github.com/MarekProkop/mminer/issues}
}

}
\keyword{internal}
